Added infinite scroll
Google and Twitter widgets have been disabled by default as the number of
that are required to load are bringing the browser to its knees especially
when scrolling. To enable them uncomment the the '@show_shares' line in
app/controllers/flickr_controller.rb

Also, this required some changes to the way the twitter and google JS files were
loaded. They have been moved to the end the page and are loaded synchronously as
they are needed in order to load their respective widgets from within the page's
JS.